Add `POST /api/v1/b/:bucket/:entry/batch` endpoint for batched writing (#355)
* refactor build.rs

* rename http_fronted -> api

* add batching write operation

* debug implementation

* fix header parsing

* clean code

* refactor implementation

* refactor error handling

* error headers

* add mutable record

* implement batching without error handling

* clean code

* reduct-rs: check errors for batched records

* GITBOOK-53: change request with no subject merged in GitBook

* update CHANGELOG, README

+++ reductstore/src/storage/block_manager.rs
@@ -13,7 +13,7 @@ use std::sync::{Arc, RwLock, Weak};
 
 use crate::storage::proto::*;
 use crate::storage::reader::RecordReader;
-use crate::storage::writer::RecordWriter;
+use crate::storage::writer::{RecordWriter, WriteChunk};
 use reduct_base::error::ReductError;
 
 pub const DEFAULT_MAX_READ_CHUNK: u64 = 1024 * 1024 * 512;
@@ -92,7 +92,7 @@ impl BlockManager {
             path,
             block,
             record_index,
-            content_length,
+            content_length as usize,
             Arc::clone(&block_manager),
         )?));
